requirements.txt
streamlit==1.28.0
requests==2.31.0
python-dotenv==1.0.0
textblob==0.17.1
python-dateutil==2.8.2